+++ src/main/java/com/readstack/crud/DiscoveryDeleter.java
@@ -1,5 +1,6 @@
 package com.readstack.crud;
 
+import com.readstack.validation.exception.DiscoveryNotFoundException;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,10 @@ class DiscoveryDeleter {
     private final DiscoveryRepository discoveryRepository;
 
     public void deleteById(Long id) {
+        if (!discoveryRepository.existsById(id)) {
+            throw new DiscoveryNotFoundException(id);
+        }
         discoveryRepository.deleteById(id);
     }
+
 }
